Add support for HDF5 files
The syntax for plotting an HDF5 file is as follows:

mpl-graph -T pdf filename:/path/to/dataset > plot.pdf

A single colon followed by a forward slash delimits the the name of the
file and the path to the dataset inside the HDF5 file. For example, to
plot the /group1/group2/my_dataset dataset in the file called
my_data.h5, one would execute:

mpl-graph -T pdf my_data.h5:/group1/group2/dataset > plot.pdf

To plot without the --auto-abscissa option, the dataset must be two
dimensional, with the first dimension being an array of abscissa, and
the second dimension being the array of line data to be plotted. If,
instead, the --auto-abscissa option is provided, the dataset must be
one-dimensional (an array) containing only the line data to be
plotted.